% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_utility_functions.R
\name{AnnotateStats}
\alias{AnnotateStats}
\title{Annotate Probe Statistics with Clustering Information}
\usage{
AnnotateStats(probe_stats, probe_clusters)
}
\arguments{
\item{probe_stats}{A dataframe containing statistics calculated by GetStats() for each probe.}

\item{probe_clusters}{A clustering object with clusters for each probe based on methylation status calculated by GetClusters().}
}
\value{
A dataframe of annotated probe statistics including clustering information.
}
\description{
This function merges probe statistics with probe annotations and clustering information,
then annotates the merged data with additional probe characteristics.
}
\keyword{internal}
