% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_utility_functions.R
\name{CountSEM_rf}
\alias{CountSEM_rf}
\title{Detect Stochastic Epigenetic Mutations (SEMs) Using Random Forest Models}
\usage{
CountSEM_rf(data, model_features, cell_counts)
}
\arguments{
\item{data}{A data frame or matrix with methylation data, samples as rows and CpG sites as columns.}

\item{model_features}{A data frame containing annotations for each feature.}

\item{cell_counts}{A data frame containing estimated cell counts for each sample by estimate.cell.proportions().}
}
\value{
A list with two elements: 'hyperSEM' and 'hypoSEM', vectors containing counts of hyper and hypo SEMs per sample.
}
\description{
Applies a Random Forest model to refine the detection of stochastic epigenetic mutations (SEMs)
identified by the IQR method, considering additional features and cell counts.
}
\note{
This function is intended for internal use and requires pre-calculated statistics and cell count estimates.
}
\keyword{internal}
