% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_blood_cells_Houseman.R
\name{estimate.cell.proportions}
\alias{estimate.cell.proportions}
\title{Estimate Cell Proportions from Methylation Data}
\usage{
estimate.cell.proportions(B, top_n = 500, mc.cores = 2, cell.coefs = NULL)
}
\arguments{
\item{B}{A matrix of beta values for methylation data.}

\item{top_n}{The number of top differentially methylated regions (DMRs) to consider.}

\item{mc.cores}{The number of cores to use for parallel processing (not used actually).}

\item{cell.coefs}{Coefficients for cell types, if pre-calculated.}
}
\value{
A matrix of estimated cell type proportions for each sample.
}
\description{
This function estimates the proportions of different cell types in a mixed tissue sample using DNA methylation data. It is based on an algorithm originally proposed by Houseman et al. (2012), adapted for use in SEMdetectR.
}
\keyword{internal}
