% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_utility_functions.R
\name{GetClusters}
\alias{GetClusters}
\title{Cluster Columns Based on Their Statistics}
\usage{
GetClusters(probe_stats)
}
\arguments{
\item{probe_stats}{A data frame of statistics for each probe, as returned by \code{GetStats}.}
}
\value{
An object of class \code{kmeans} representing the clustering result.
}
\description{
Performs k-means clustering on the scaled statistics of probes in DNA methylation dataframe,
aiming to group columns into 3 clusters based on their statistical properties.
}
\note{
This function is intended for internal use within the package.
}
\keyword{internal}
