% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_utility_functions.R
\name{GetStats}
\alias{GetStats}
\title{Calculate Statistics for Each probe in a DNA methylation dataframe}
\usage{
GetStats(betas, num_cores = 1)
}
\arguments{
\item{betas}{A data frame or matrix of numerical data.}

\item{num_cores}{The number of cores to use for parallel computation.}
}
\value{
A data frame with computed statistics for each column.
}
\description{
Computes mean, quartiles, skewness, standard deviation, minimum, maximum, and kurtosis.
Can operate in parallel to speed up computations.
}
\note{
This function is intended for internal use within the package.
}
\keyword{internal}
