% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseman_dmrs_locs.R
\docType{data}
\name{houseman_dmrs_locs.rda}
\alias{houseman_dmrs_locs.rda}
\alias{houseman_dmrs_locs}
\title{coefficients used by celltypes450 to infer cell types}
\format{
A data frame with 473 rows and 8 columns;
}
\source{
Extracted from the \code{celltypes450} package.
}
\usage{
houseman_dmrs_locs
}
\description{
Coefficients for cell count inference from Houseman et al.
}
\keyword{datasets}
