% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe_annotations.R
\docType{data}
\name{probe_annotations.rda}
\alias{probe_annotations.rda}
\alias{probe_annotations}
\title{Illumina Probe Annotations Dataframe}
\format{
A data frame with 485512 rows and 9 columns:
\describe{
\item{\code{Name}}{Name of Illumina probe.}
\item{\code{Color}}{Color of Illumina probe.}
\item{\code{SourceSeq}}{Nucleotide sequence.}
\item{\code{SBE_rs}}{SNPs at single-base extension.}
\item{\code{DMR}}{Differentially methylated region.}
\item{\code{Enhancer}}{Enhancer.}
\item{\code{regFeature}}{Regulatory features.}
\item{\code{DHS}}{DNase I hypersensitive sites.}
\item{\code{location}}{Genomic location.}
}
}
\source{
Taken from the Illumina manifest for the 450k array
}
\usage{
probe_annotations
}
\description{
A dataframe containing probe annotations used as features in the Random Forest models.
Includes information on Name, Color, Source Sequence, presence of SNP at SBE, overlap with a differentially methylated region, Enhancer, regulatory features, DNase I hypersensitive sites, and type of genomic location of the probe.
}
\keyword{datasets}
